Adapt measurements downloading to Sessionable protocol

+++ AirCasting/Services/DownloadMeasurmentsService.swift
@@ -18,26 +18,6 @@ protocol MeasurementUpdatingService {
 }
 
 final class DownloadMeasurementsService: MeasurementUpdatingService {
-    enum Session {
-        case session(SessionEntity)
-        case externalSession(ExternalSessionEntity)
-        
-        var uuid: String {
-            switch self {
-            case .session(let sessionEntity):
-                return sessionEntity.uuid.rawValue
-            case .externalSession(let externalSessionEntity):
-                return externalSessionEntity.uuid
-            }
-        }
-        
-        enum SessionType {
-            case regular
-            case external
-        }
-    }
-    
-    
     @Injected private var persistenceController: PersistenceController
     private let fixedSessionService = FixedSessionAPIService()
     private var timerSink: Cancellable?
@@ -56,28 +36,28 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
     
     func downloadMeasurements(for sessionUUID: SessionUUID, lastSynced: Date, completion: @escaping () -> Void) {
         lastFetchCancellableTask = fixedSessionService.getFixedMeasurement(uuid: sessionUUID, lastSync: lastSynced) { [weak self] in
-            self?.processServiceResponse($0, for: sessionUUID, type: .regular, completion: completion)
+            self?.processServiceResponse($0, for: sessionUUID, isExternal: false, completion: completion)
         }
     }
     
-    private func updateMeasurements(for sessionUUID: SessionUUID, lastSynced: Date, type: Session.SessionType) {
+    private func updateMeasurements(for sessionUUID: SessionUUID, lastSynced: Date, isExternal: Bool) {
         lastFetchCancellableTask = fixedSessionService.getFixedMeasurement(uuid: sessionUUID, lastSync: lastSynced) { [weak self] in
-            self?.processServiceResponse($0, for: sessionUUID, type: type)
+            self?.processServiceResponse($0, for: sessionUUID, isExternal: isExternal)
         }
     }
     
     func updateAllSessionsMeasurements() {
         getAllSessionsData() { [unowned self] sessionsData in
             Log.info("Scheduled measurements update triggered (session count: \(sessionsData.count))")
-            sessionsData.forEach { self.updateMeasurements(for: $0.uuid, lastSynced: $0.lastSynced, type: $0.type) }
+            sessionsData.forEach { self.updateMeasurements(for: $0.uuid, lastSynced: $0.lastSynced, isExternal: $0.isExternal) }
         }
     }
     
-    private func getAllSessionsData(completion: @escaping ([(uuid: SessionUUID, lastSynced: Date, type: Session.SessionType)]) -> Void) {
+    private func getAllSessionsData(completion: @escaping ([(uuid: SessionUUID, lastSynced: Date, isExternal: Bool)]) -> Void) {
         let request: NSFetchRequest<SessionEntity> = SessionEntity.fetchRequest()
         request.predicate = NSPredicate(format: "followedAt != NULL")
         let context = persistenceController.editContext
-        var returnData: [(uuid: SessionUUID, lastSynced: Date, type: Session.SessionType)] = []
+        var returnData: [(uuid: SessionUUID, lastSynced: Date, isExternal: Bool)] = []
         
         let externalSessionsRequest = ExternalSessionEntity.fetchRequest()
         request.predicate = NSPredicate(value: true)
@@ -86,8 +66,8 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
             do {
                 let sessions = try context.fetch(request)
                 let externalSessions = try context.fetch(externalSessionsRequest)
-                let mappedSessions = sessions.map { ($0.uuid!, self.getSyncDate(for: $0), Session.SessionType.regular) }
-                let mappedExternalSessions = externalSessions.map { (SessionUUID(uuidString: $0.uuid)!, self.getExternalSessionSyncDate(for: $0), Session.SessionType.external) }
+                let mappedSessions = sessions.map { ($0.uuid!, self.getSyncDate(for: $0), $0.isExternal) }
+                let mappedExternalSessions = externalSessions.map { ($0.uuid!, self.getExternalSessionSyncDate(for: $0), $0.isExternal) }
                 returnData = mappedSessions + mappedExternalSessions
                 completion(returnData)
             } catch {
@@ -106,7 +86,7 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
     }
     
     private func getExternalSessionSyncDate(for session: ExternalSessionEntity?) -> Date {
-        let lastMeasurementTime = session?.measurementStreams
+        let lastMeasurementTime = session?.allStreams
             .compactMap(\.lastMeasurementTime)
             .sorted()
             .last
@@ -115,10 +95,10 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
     }
     
     private func processServiceResponse(_ response: Result<FixedSession.FixedMeasurementOutput, Error>,
-                                        for sessionUUID: SessionUUID, type: Session.SessionType, completion: () -> Void = {}) {
+                                        for sessionUUID: SessionUUID, isExternal: Bool, completion: () -> Void = {}) {
         switch response {
         case .success(let response):
-            processServiceOutput(response, for: sessionUUID, type: type)
+            processServiceOutput(response, for: sessionUUID, isExternal: isExternal)
             completion()
         case .failure(let error):
             Log.warning("Failed to fetch measurements for uuid '\(sessionUUID)' \(error)")
@@ -126,24 +106,25 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
     }
     
     private func processServiceOutput(_ output: FixedSession.FixedMeasurementOutput,
-                                      for sessionUUID: SessionUUID, type: Session.SessionType) {
+                                      for sessionUUID: SessionUUID,
+                                      isExternal: Bool) {
         Log.info("Processing download measurements response for: \(sessionUUID)")
         let context = persistenceController.editContext
         context.perform {
             do {
-                switch type {
-                case .regular:
+                if !isExternal {
                     Log.info("Processing regular session response")
                     let session: SessionEntity = try context.newOrExisting(uuid: output.uuid)
                     try UpdateSessionParamsService().updateSessionsParams(session: session, output: output)
                     try self.removeOldService.removeOldestMeasurements(in: context,
-                                                                       from: sessionUUID, of: type)
-                case .external:
+                                                                       from: sessionUUID,
+                                                                       which: isExternal)
+                } else {
                     Log.info("Processing external session response")
-                    let session = try context.existingExternalSession(uuid: sessionUUID.rawValue)
+                    let session = try context.existingExternalSession(uuid: sessionUUID)
                     session.endTime = output.end_time
                     output.streams.forEach({ stream in
-                        if let sessionStream = session.measurementStreams.first(where: { $0.sensorName == stream.key }) {
+                        if let sessionStream = session.allStreams.first(where: { $0.sensorName == stream.key }) {
                             stream.value.measurements.forEach({ measurement in
                                 let newMeasurement = MeasurementEntity(context: context)
                                 newMeasurement.location = CLLocationCoordinate2D(latitude: measurement.latitude, longitude: measurement.longitude)
@@ -154,7 +135,8 @@ final class DownloadMeasurementsService: MeasurementUpdatingService {
                         }
                     })
                     try self.removeOldService.removeOldestMeasurements(in: context,
-                                                                       from: sessionUUID, of: type)
+                                                                       from: sessionUUID,
+                                                                       which: isExternal)
                 }
                 try context.save()
             } catch let error as UpdateSessionParamsService.Error {
